+++ mlos_bench/mlos_bench/environments/__init__.py
@@ -12,7 +12,7 @@ from mlos_bench.environments.base_environment import Environment
 from mlos_bench.environments.mock_env import MockEnv
 from mlos_bench.environments.remote.remote_env import RemoteEnv
 from mlos_bench.environments.local.local_env import LocalEnv
-from mlos_bench.environments.local.local_env_fileshare import LocalFileShareEnv
+from mlos_bench.environments.local.local_fileshare_env import LocalFileShareEnv
 from mlos_bench.environments.composite_env import CompositeEnv
 
 __all__ = [
